Rename `_helpers.py` to `helpers.py`:
No need to make helper modules private as they are already located in a
private subpackage (`_2d` and `_3d`, respectively).
