+++ optbeam/_2d/helpers.py
