+++ optbeam/_3d/helpers.py
